requirements.txt
fastapi==0.104.1
sqlmodel==0.0.8
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pytest==7.4.3
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai>=1.0.0
mcp>=1.0.0
google-generativeai>=0.3.0
slowapi==0.1.9